requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.4
Brlapi==0.8.3
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.7
DateTime==5.5
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.7
dotenv==0.9.9
duckduckgo_search==7.5.2
duplicity==0.8.21
exceptiongroup==1.2.2
fastapi==0.115.11
fasteners==0.14.1
filelock==3.18.0
folder-color==0.2.0
frozenlist==1.5.0
fsspec==2025.3.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gritql==0.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
icoextract==0.1.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
langchain==0.3.21
langchain-cli==0.0.36
langchain-community==0.3.20
langchain-core==0.3.46
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.7
langserve==0.3.1
langsmith==0.3.18
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.44.0
lockfile==0.12.2
louis==3.20.0
lxml==5.3.1
macaroonbakery==1.3.1
magic-filter==1.0.12
Mako==1.1.3
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.4.2
ffmpeg-python==0.2.0
numba==0.61.0
numpy==2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai-whisper==20240930
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pefile==2021.9.3
pexpect==4.8.0
pillow==11.0.0
primp==0.14.0
proglog==0.1.10
propcache==0.3.0
protobuf==3.12.4
psutil==5.9.0
psycopg==3.2.6
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyelftools==0.27
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==5.4.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt===2.4.0-ubuntu4-zorin1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2025.1
pyxdg==0.27
PyYAML==6.0.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
SecretStorage==3.3.1
selenium==4.29.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
smolagents==1.11.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.5.3
sse-starlette==1.8.2
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
wadllib==1.3.6
webdriver-manager==4.0.2
websocket-client==1.8.0
whisper==1.1.10
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
zope.interface==7.2
zorin-appearance==5.3
zorin-connect==1.2
zorin-exec-guard==1.5
zstandard==0.23.0
